+++ .github/consumer/Program.cs
@@ -1,11 +1,14 @@
 using System;
 using System.Collections.Generic;
 using Arlecchino.Atoms;
+using Arlecchino.Atoms.Local;
+using Arlecchino.Atoms.Tracked;
 using Arlecchino.Commands;
 using Arlecchino.Hosting;
 using Arlecchino.Input;
 using Arlecchino.Navigation;
 using Arlecchino.Rendering;
+using Arlecchino.Rendering.Colors;
 using Arlecchino.Widgets;
 using Consumer.Navigation;
 using Microsoft.Extensions.DependencyInjection;